- Removed Husky. - Update seeding data & .nginx & deploy.sh. - Formatting.

+++ deploy.sh
@@ -10,10 +10,22 @@ PATH=$PATH:/home/$USER/.nvm/versions/node/lts/bin
 # Stop portfolio service
 pm2 delete nextjs-portfolio && pm2 save -f
 
-# Copy database file
+# Copy database file if exists
 if [ -f "$BASE_PATH/prisma/db.sqlite" ]
 then
         cp -f "$BASE_PATH/prisma/db.sqlite" "$SCRIPT_PATH/db.sqlite"
+fi
+
+# Copy .env file if exists
+if [ -f "$BASE_PATH/.env" ]
+then
+        cp -f "$BASE_PATH/.env" "$SCRIPT_PATH/.env"
+fi
+
+# Copy config file if exists
+if [ -f "$BASE_PATH/src/app/config.ts" ]
+then
+        cp -f "$BASE_PATH/src/app/config.ts" "$SCRIPT_PATH/config.ts"
 fi
 
 
@@ -21,17 +33,25 @@ fi
 rm -rf $BASE_PATH/* $BASE_PATH/.*
 
 # Clone Repository
-git clone https://github.com/AlanD20/aland20.tech.git "$BASE_PATH/."
+git clone https://github.com/AlanD20/aland20.com.git "$BASE_PATH/."
 
-# Copy env file
-cp -f "$SCRIPT_PATH/.env" "$BASE_PATH/.env"
+# Recover .env file if exists
+if [ -f "$SCRIPT_PATH/.env" ]
+then
+        cp -f "$SCRIPT_PATH/.env" "$BASE_PATH/.env"
+fi
 
 # Install dependencies
 cd "$BASE_PATH" && yarn && yarn db:reset
 
-# Setup Project
-cp -f "$SCRIPT_PATH/config.ts" "$BASE_PATH/src/app/config.ts"
+# Recover Config file if exists
+if [ -f "$SCRIPT_PATH/config.ts" ]
+then
+        cp -f "$SCRIPT_PATH/config.ts" "$BASE_PATH/src/app/config.ts"
+fi
 
+
+# Recover database if exists
 if [ -f "$SCRIPT_PATH/db.sqlite" ]
 then
         cp -f "$SCRIPT_PATH/db.sqlite" "$BASE_PATH/prisma/db.sqlite"
@@ -53,7 +73,7 @@ else
         echo "- [x] Failed to restore database!"
 fi
 
-if grep -Fq "https://aland20.tech" "$BASE_PATH/src/app/config.ts"
+if grep -Fq "https://aland20.com" "$BASE_PATH/src/app/config.ts"
 then
         echo "- [x] Config file updated!"
 else
